# Completion benchmark