--- a/.gitignore
+++ b/.gitignore
@@ -155,3 +155,6 @@ yarn-*.log
 
 # Bundled extension assets
 *.vsix
+
+# testapp modifications
+src/Razor/test/testapps/ComponentApp/Components/Pages/Generated.razor