--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptor.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptor.cs
@@ -6,6 +6,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+
 using Microsoft.AspNetCore.Razor.Language.Components;
 
 namespace Microsoft.AspNetCore.Razor.Language;
@@ -50,21 +51,23 @@ public abstract class BoundAttributeDescriptor : IEquatable<BoundAttributeDescri
 
     public bool CaseSensitive { get; protected set; }
 
-    private bool _directiveAttributeSet;
-    private bool _isDirectiveAttribute;
+    // We need this to be atomic so:
+    // 0: Uninitialized
+    // 1: false
+    // 2: true
+    private int _isDirectiveAttribute = 0;
 
     public bool IsDirectiveAttribute
     {
         get
         {
-            if (!_directiveAttributeSet)
+            if (_isDirectiveAttribute == 0)
             {
                 _isDirectiveAttribute = Metadata.TryGetValue(ComponentMetadata.Common.DirectiveAttribute, out var value) &&
-                        string.Equals(bool.TrueString, value);
-                _directiveAttributeSet = true;
+                        string.Equals(bool.TrueString, value) ? 2 : 1;
             }
 
-            return _isDirectiveAttribute;
+            return _isDirectiveAttribute == 2;
         }
     }