--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptorExtensions.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/BoundAttributeDescriptorExtensions.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #nullable disable
@@ -66,14 +66,12 @@ public static class BoundAttributeDescriptorExtensions
 
     public static bool IsDirectiveAttribute(this BoundAttributeDescriptor attribute)
     {
-        if (attribute == null)
+        if (attribute is null)
         {
             throw new ArgumentNullException(nameof(attribute));
         }
 
-        return
-            attribute.Metadata.TryGetValue(ComponentMetadata.Common.DirectiveAttribute, out var value) &&
-            string.Equals(bool.TrueString, value);
+        return attribute.IsDirectiveAttribute;
     }
 
     public static bool IsDefaultKind(this BoundAttributeParameterDescriptor parameter)