--- a/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagHelperBinder.cs
+++ b/src/Compiler/Microsoft.AspNetCore.Razor.Language/src/TagHelperBinder.cs
@@ -1,4 +1,4 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
+// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #nullable disable
@@ -132,6 +132,8 @@ internal sealed class TagHelperBinder
 
     private void Register(TagHelperDescriptor descriptor)
     {
+        bool descriptorAdded = false;
+
         var count = descriptor.TagMatchingRules.Count;
         for (var i = 0; i < count; i++)
         {
@@ -148,7 +150,12 @@ internal sealed class TagHelperBinder
                 _registrations[registrationKey] = descriptorSet;
             }
 
-            descriptorSet.Add(descriptor);
+            // We don't need to keep trying to add the same Descriptor to the HashSet for each matching rule.
+            if (!descriptorAdded)
+            {
+                _ = descriptorSet.Add(descriptor);
+                descriptorAdded = true;
+            }
         }
     }
 }