--- a/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/ProjectSystem/ProjectSnapshotManagerBenchmarkBase.cs
+++ b/src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/ProjectSystem/ProjectSnapshotManagerBenchmarkBase.cs
@@ -24,6 +24,7 @@ public abstract partial class ProjectSnapshotManagerBenchmarkBase
     internal ImmutableArray<HostDocument> Documents { get; }
     internal ImmutableArray<TextLoader> TextLoaders { get; }
     internal TagHelperResolver TagHelperResolver { get; }
+    protected string RepoRoot { get; }
 
     protected ProjectSnapshotManagerBenchmarkBase()
     {
@@ -33,8 +34,8 @@ public abstract partial class ProjectSnapshotManagerBenchmarkBase
             current = current.Parent;
         }
 
-        var root = current ?? throw new InvalidOperationException("Could not find Razor.sln");
-        var projectRoot = Path.Combine(root.FullName, "src", "Razor", "test", "testapps", "LargeProject");
+        RepoRoot = current?.FullName ?? throw new InvalidOperationException("Could not find Razor.sln");
+        var projectRoot = Path.Combine(RepoRoot, "src", "Razor", "test", "testapps", "LargeProject");
 
         HostProject = new HostProject(Path.Combine(projectRoot, "LargeProject.csproj"), FallbackRazorConfiguration.MVC_2_1, rootNamespace: null);
 
@@ -62,8 +63,13 @@ public abstract partial class ProjectSnapshotManagerBenchmarkBase
 
         Documents = documents.ToImmutable();
 
-        var tagHelpers = Path.Combine(root.FullName, "src", "Razor", "benchmarks", "Microsoft.AspNetCore.Razor.Microbenchmarks", "taghelpers.json");
-        TagHelperResolver = new StaticTagHelperResolver(ReadTagHelpers(tagHelpers), NoOpTelemetryReporter.Instance);
+        TagHelperResolver = new StaticTagHelperResolver(GetTagHelperDescriptors(), NoOpTelemetryReporter.Instance);
+    }
+
+    internal IReadOnlyList<TagHelperDescriptor> GetTagHelperDescriptors()
+    {
+        var tagHelpers = Path.Combine(RepoRoot, "src", "Razor", "benchmarks", "Microsoft.AspNetCore.Razor.Microbenchmarks", "taghelpers.json");
+        return ReadTagHelpers(tagHelpers);
     }
 
     internal DefaultProjectSnapshotManager CreateProjectSnapshotManager()