--- a/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/CompletionListProvider.cs
+++ b/src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Completion/CompletionListProvider.cs
@@ -41,7 +41,7 @@ internal class CompletionListProvider
             : null;
 
         // Extract the items we got back from the delegated server, to inform tag helper completion
-        var existingItems = delegatedCompletionList != null
+        var existingItems = delegatedCompletionList?.Items != null
             ? new HashSet<string>(delegatedCompletionList.Items.Select(i => i.Label))
             : null;