--- a/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/LanguageServerTagHelperCompletionService.cs
+++ b/src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/LanguageServerTagHelperCompletionService.cs
@@ -242,7 +242,7 @@ internal class LanguageServerTagHelperCompletionService : TagHelperCompletionSer
                 var completionTagName = kvp.Key;
                 var tagHelperDescriptors = kvp.Value;
 
-                if (tagHelperDescriptors.Any() ||
+                if (tagHelperDescriptors.Count > 0 ||
                     (!string.IsNullOrEmpty(prefix) && completionTagName.StartsWith(prefix, StringComparison.OrdinalIgnoreCase)))
                 {
                     // The current completion either has other TagHelper's associated with it or is prefixed with a non-empty