… src/Razor/test/testapps/ComponentApp/Components/Pages/Generated.razor