﻿## Summary of the changes

- Created the Completion Benchmarks and made sure they were doing some real work.
- Made some experimental changes to see how we could move the needle on the results. I'm not convinced these are worth the complexity they add to the code/churn in shared areas, but I thought I'd put them out there to let people know what I had already tried. If people agree with me I'll go ahead and delete them.

## Results

### Before Experiments

|             Method |        Job |            Toolchain |     Mean |     Error |    StdDev |   Median |      Min |      Max |     Gen0 |     Gen1 |    Gen2 | Allocated |
|------------------- |----------- |--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|----------:|
| 'Razor Completion' | Job-EJEXYW |             .NET 7.0 | 6.367 ms | 0.1106 ms | 0.1035 ms | 6.360 ms | 6.196 ms | 6.569 ms | 398.4375 | 101.5625 | 23.4375 |   2.29 MB |
| 'Razor Completion' | Job-DSHCRS | .NET Framework 4.7.2 | 7.897 ms | 0.1555 ms | 0.1791 ms | 7.927 ms | 7.605 ms | 8.243 ms | 414.0625 |  93.7500 | 23.4375 |   2.36 MB |

### After Experiments

|             Method |        Job |            Toolchain |     Mean |     Error |    StdDev |   Median |      Min |      Max |     Gen0 |     Gen1 |    Gen2 | Allocated |
|------------------- |----------- |--------------------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|----------:|
| 'Razor Completion' | Job-EJEXYW |             .NET 7.0 | 4.195 ms | 0.0830 ms | 0.1019 ms | 4.193 ms | 4.027 ms | 4.428 ms | 398.4375 | 101.5625 | 23.4375 |   2.29 MB |
| 'Razor Completion' | Job-DSHCRS | .NET Framework 4.7.2 | 5.827 ms | 0.0984 ms | 0.0920 ms | 5.807 ms | 5.692 ms | 5.997 ms | 671.8750 | 125.0000 | 31.2500 |   3.88 MB |

As you can see there was a reasonable improvement in speed as a percentile, but when we're operating on numbers this small I have to wonder how worthwhile the changes actually are. My opinion is that the work @DustinCampbell is doing WRT making Roslyn faster in this scenario will dwarf these changes from a user perspective both in terms of memory and CPU time so we're better off without the complications they add for the relatively small benefits of avoiding a couple Dictionary/HashSet lookups. The one exception IMO would be the `.Count > 0 ` to `.Any()` changes since I feel they're just more clear.

Tiny part of https://github.com/dotnet/razor/issues/8111